Turn ethernet into class; standard interface like WSGI
WSGI uses functions of the form:

```
    def app(environ, start_response):
        ...
```

For detour, functions take the form:

```
    def filter(recived_data, write_back, write_fwd):
        ...
```

The function itself does not inerhently care which side (Alice or Bob)
of the connection it is on. This system is also very easy to build
pre-processors for in the form of decorators. (i.e. a @HTTP decorator
which only works at the application layer).